+++ paddle/pten/include/manipulation.h
@@ -20,7 +20,7 @@
 #include "paddle/pten/kernels/cpu/manipulation.h"
 #include "paddle/pten/kernels/flatten_kernel.h"
 #include "paddle/pten/kernels/gpu/manipulation.h"
-#include "paddle/pten/kernels/xpu/manipulation.h"
+#include "paddle/pten/kernels/reshape_kernel.h"
 
 namespace pten {
 
@@ -61,7 +61,7 @@ DenseTensor Reshape(const ContextT& dev_ctx,
       pten::make_intrusive<paddle::experimental::SharedStorage>(
           dev_ctx.GetPlace()),
       std::move(out_meta));
-  Reshape(dev_ctx, x, ScalarArray(shape), &dense_out);
+  Reshape<ContextT>(dev_ctx, x, ScalarArray(shape), &dense_out);
   return dense_out;
 }
