+++ paddle/pten/kernels/reshape_kernel.h
@@ -14,26 +14,22 @@ limitations under the License. */
 
 #pragma once
 
-#ifdef PADDLE_WITH_XPU
-
-#include "paddle/pten/backends/xpu/xpu_context.h"
 #include "paddle/pten/common/scalar_array.h"
 #include "paddle/pten/core/dense_tensor.h"
-#include "paddle/pten/core/kernel_registry.h"
 
 namespace pten {
 
-void Reshape(const XPUContext& dev_ctx,
+template <typename ContextT>
+void Reshape(const ContextT& dev_ctx,
              const DenseTensor& x,
              const ScalarArray& shape,
              DenseTensor* out);
 
-void ReshapeWithXShape(const XPUContext& dev_ctx,
+template <typename ContextT>
+void ReshapeWithXShape(const ContextT& dev_ctx,
                        const DenseTensor& x,
                        const ScalarArray& shape,
                        DenseTensor* xshape,
                        DenseTensor* out);
 
 }  // namespace pten
-
-#endif
